Reorg skill_entry_from_url into OSM class Update unit tests
