+++ ovos_skills_manager/github/utils.py
@@ -169,24 +169,3 @@ def match_url_template(url, template, branch=None):
         return blob2raw(url)
     raise GithubInvalidUrl
 
-
-def get_skill_entry_from_url(url: str):
-    from ovos_skills_manager.github import get_branch_from_github_url, normalize_github_url, get_requirements_json, \
-        get_skill_json
-    from ovos_skills_manager.skill_entry import SkillEntry
-    try:
-        branch = get_branch_from_github_url(url)
-    except GithubInvalidBranch:
-        branch = None
-    url = normalize_github_url(url)
-    requirements = get_requirements_json(url, branch)
-    requirements["system"] = {k: v.split() for k, v in requirements.get("system", {}).items()}
-    try:
-        json = get_skill_json(url, branch)
-        requirements = merge_dict(requirements, json.get("requirements", {}),
-                                  merge_lists=True, skip_empty=True, no_dupes=True)
-    except GithubFileNotFound:
-        pass
-    return SkillEntry.from_json({"url": url,
-                                 "branch": branch,
-                                 "requirements": requirements}, False)
